Fixed a few compilation errors
