+++ include/common/rng.h
@@ -1,10 +1,37 @@
-#ifndef DISTRIBUTIONS_H
-#define DISTRIBUTIONS_H
+#ifndef RNG_H
+#define RNG_H
 
 #include <stdint.h>
-#include "utils.h"
-#include "structure_alias.h"
+#include "maths_structures.h"
 #include "spqlios_alias.h"
+
+// Random Numbers
+
+/**
+ * Generates an uniform random number.
+ *
+ * @param result A pointer that will point to the generated value.
+ * @param nb_bits The number of bits of the result, sign bit included.
+ * 
+ * @retval `-1` if an error occurs. A message will be printed in this case.
+ * @retval `0` otherwise.
+ */
+int rand_uniform(int64_t *result, uint64_t nb_bits);
+
+/**
+ * Generates a gaussian random number with the given parameters.
+ *
+ * @param result A pointer that will point to the generated value.
+ * @param mu     The mean value.
+ * @param sigma  The standard deviation.
+ *
+ * @retval `-1` if an error occurs. A message will be printed in this case.
+ * @retval `0` otherwise.
+ */
+int rand_normal(double *result, double mu, double sigma);
+
+
+// Random Vectors
 
 /**
  * @brief Generates a random vector following a uniform distribution.
@@ -58,4 +85,4 @@ int new_uniform_random_vec_dft(MODULE* module, PolyUnivDFT* res_dft, int64_t res
  */
 int new_normal_random_vec(uint64_t limb_len, double* res, int64_t res_size, int64_t res_sl, double mu, double sigma);
 
-#endif // DISTRIBUTION_H
+#endif // RNG_H
