+++ src/common/spqlios_alias.c
@@ -1,7 +1,5 @@
 #include "spqlios_alias.h"
 
-#include "vec_znx_arithmetic_private.h"
-
 MODULE* new_module_info_p(uint64_t N) {
   return new_module_info(N, FFT64);
 }
